requirements.txt
# List of required packages (Dependencies)
python-telegram-bot
numpy

